Add Renamer class
