+++ MovieRenamer/MovieRenamer.csproj
@@ -61,6 +61,7 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Renamer.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
